+++ app/components/Links/default.tsx
@@ -1,3 +1,5 @@
+import { cn } from "~/lib/utils";
+
 import { RouteLinkProps } from ".";
 import { RouteButton } from "../RouteButton";
 
@@ -5,7 +7,13 @@ type DefaultRouteLinkProps = RouteLinkProps;
 
 export const DefaultLink = ({ label, ...props }: DefaultRouteLinkProps) => {
   return (
-    <RouteButton className="bg-background " {...props}>
+    <RouteButton
+      className={cn(
+        "bg-background text-primary/50"
+        // !props.disabled && "hover:text-primary/80"
+      )}
+      {...props}
+    >
       {label ?? "click"}
     </RouteButton>
   );