+++ app/components/Links/shop.tsx
@@ -1,3 +1,5 @@
+import { cn } from "~/lib/utils";
+
 import { RouteLinkProps } from ".";
 import { ShopIcon } from "../Icons/ShopIcon";
 import { RouteButton } from "../RouteButton";
@@ -8,7 +10,11 @@ export const ShopLink = ({ label, ...props }: ShopRouteLinkProps) => {
   return (
     <RouteButton
       icon={ShopIcon}
-      className="border border-primary bg-primary text-white dark:hover:bg-primary/10 dark:bg-secondary/10 "
+      className={cn(
+        "border border-primary bg-primary text-white ",
+        !props.disabled &&
+          "hover:bg-primary/80  dark:hover:bg-primary/10 dark:bg-secondary/10"
+      )}
       {...props}
     >
       {label ?? "Shop"}