+++ app/components/RouteButton/index.tsx
@@ -3,12 +3,14 @@ import { cn } from "~/lib/utils";
 
 interface RouteButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   icon?: React.ElementType;
+  disabled?: boolean;
 }
 
 export const RouteButton = ({
   icon: Icon,
   className,
   children,
+  disabled = false,
   ...props
 }: RouteButtonProps) => {
   return (
@@ -16,9 +18,11 @@ export const RouteButton = ({
       className={cn(
         "flex justify-center items-center gap-2",
         "transition-colors duration-300",
-        "w-full px-4 py-4 rounded-sm",
+        "w-full px-4 py-3 rounded-sm",
+        "text-sm",
         // "hover:scale-105",
-        className
+        className,
+        disabled && "opacity-20 cursor-not-allowed"
       )}
       {...props}
     >