+++ app/root.tsx
@@ -1,6 +1,7 @@
 import type { LinksFunction } from "@remix-run/cloudflare";
 import {
   Links,
+  LiveReload,
   Meta,
   Outlet,
   Scripts,
@@ -43,6 +44,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
         <div className="max-w-lg mx-auto h-screen px-4">{children}</div>
         <ScrollRestoration />
         <Scripts />
+        {process.env.NODE_ENV === "development" && <LiveReload />}
       </body>
     </html>
   );