+++ app/routes/_index.tsx
@@ -1,6 +1,4 @@
 import { MetaFunction } from "@remix-run/cloudflare";
-import { TopFixedRight } from "~/components/Fixed/TopRight";
-import { ToggleTheme } from "~/components/Toggle/Theme";
 import { config } from "~/config";
 
 import Linktr from "./linktr";
@@ -18,14 +16,7 @@ export const meta: MetaFunction = () => {
 };
 
 const Index = () => {
-  return (
-    <>
-      <Linktr />
-      <TopFixedRight>
-        <ToggleTheme />
-      </TopFixedRight>
-    </>
-  );
+  return <Linktr />;
 };
 
 export default Index;